+++ javascript/modules/client/src/context.ts
@@ -1,6 +1,7 @@
 import { ContextState, ContextParams } from "./internal/interfaces/context";
 import { JsonRpcProvider, Networkish } from "@ethersproject/providers";
 import { UnsupportedProtocolError } from "@aragon/sdk-common";
+import { activeContractsList } from "@aragon/core-contracts-ethers";
 // import { create as ipfsCreate, Options as IpfsOptions } from "ipfs-http-client";
 // import { GraphQLClient } from "graphql-request";
 export { ContextParams } from "./internal/interfaces/context";
@@ -16,7 +17,6 @@ if (typeof process !== "undefined" && process.env?.TESTING) {
 const defaultState: ContextState = {
   network: "mainnet",
   dao: "",
-  daoFactoryAddress: "",
   web3Providers: [],
   gasFeeEstimationFactor: DEFAULT_GAS_FEE_ESTIMATION_FACTOR,
 };
@@ -91,6 +91,11 @@ export class Context {
     }
     if (contextParams.daoFactoryAddress) {
       this.state.daoFactoryAddress = contextParams.daoFactoryAddress;
+    } else if (this.state.network.toString() in activeContractsList) {
+      this.state.daoFactoryAddress =
+        activeContractsList[
+          this.state.network.toString() as keyof typeof activeContractsList
+        ].DAOFactory;
     }
     if (contextParams.signer) {
       this.state.signer = contextParams.signer;
@@ -203,8 +208,8 @@ export class Context {
    *
    * @public
    */
-  get daoFactoryAddress(): string {
-    return this.state.daoFactoryAddress || defaultState.daoFactoryAddress;
+  get daoFactoryAddress(): string | undefined {
+    return this.state.daoFactoryAddress;
   }
 
   /**